myfittempoapp/requirements.txt
django
pymysql
djangorestframework
Pillow
firebase-admin
django-cors-headers
# pycrypto